[0.3.7] rewrite AssetManager to own assets and give everyone references to his data | update project structure

+++ src/ktanks/core/constants/TerrainData.h
@@ -1,9 +1,11 @@
 #ifndef KTANKS_TERRAINSPRITELIST_H
 #define KTANKS_TERRAINSPRITELIST_H
 
+#include <cstdint>
+
 namespace ktanks {
 
-    enum class TerrainSprite : int {
+    enum class TerrainSprite : uint8_t {
         Grass1,
         Grass2,
         GCornerLL,
